Use new canary and add more stories
