+++ app/note/edit/page.stories.tsx
@@ -1,0 +1,38 @@
+import { Meta, StoryObj } from '@storybook/react'
+import { cookies } from 'next/headers'
+import Page from './page'
+import { prisma } from '#lib/db'
+import { createUserCookie, userCookieKey } from '#lib/session'
+import { PageDecorator } from '#.storybook/decorators'
+
+const meta = {
+  component: Page,
+  parameters: { layout: 'fullscreen' },
+  decorators: [PageDecorator],
+  async beforeEach() {
+    cookies().set(userCookieKey, await createUserCookie('storybookjs'))
+
+    await prisma.note.create({
+      data: {
+        id: '1',
+        title: 'Module mocking in Storybook?',
+        body: "Yup, that's a thing now! 🎉",
+        createdBy: 'storybookjs',
+      },
+    })
+    await prisma.note.create({
+      data: {
+        id: '2',
+        title: 'RSC support as well??',
+        body: 'RSC is pretty cool, even cooler that Storybook supports it!',
+        createdBy: 'storybookjs',
+      },
+    })
+  },
+} satisfies Meta<typeof Page>
+
+export default meta
+
+type Story = StoryObj<typeof meta>
+
+export const Default: Story = {}
