+++ package.json
@@ -50,13 +50,13 @@
   "devDependencies": {
     "@chromatic-com/storybook": "^1.2.25",
     "@prisma/client": "^5.11.0",
-    "@storybook/addon-actions": "0.0.0-pr-26853-sha-8dbd04c4",
-    "@storybook/addon-essentials": "0.0.0-pr-26853-sha-8dbd04c4",
-    "@storybook/addon-interactions": "0.0.0-pr-26853-sha-8dbd04c4",
-    "@storybook/blocks": "0.0.0-pr-26853-sha-8dbd04c4",
-    "@storybook/nextjs": "0.0.0-pr-26853-sha-8dbd04c4",
-    "@storybook/react": "0.0.0-pr-26853-sha-8dbd04c4",
-    "@storybook/test": "0.0.0-pr-26853-sha-8dbd04c4",
+    "@storybook/addon-actions": "0.0.0-pr-26886-sha-b094da7b",
+    "@storybook/addon-essentials": "0.0.0-pr-26886-sha-b094da7b",
+    "@storybook/addon-interactions": "0.0.0-pr-26886-sha-b094da7b",
+    "@storybook/blocks": "0.0.0-pr-26886-sha-b094da7b",
+    "@storybook/nextjs": "0.0.0-pr-26886-sha-b094da7b",
+    "@storybook/react": "0.0.0-pr-26886-sha-b094da7b",
+    "@storybook/test": "0.0.0-pr-26886-sha-b094da7b",
     "@storybook/test-runner": "^0.17.0",
     "@types/node": "^20",
     "@types/react": "^18",
@@ -73,7 +73,7 @@
     "prettier": "^3.2.5",
     "prisma": "^5.11.0",
     "prisma-mock": "^0.10.0",
-    "storybook": "0.0.0-pr-26853-sha-8dbd04c4",
+    "storybook": "0.0.0-pr-26886-sha-b094da7b",
     "ts-node": "^10.9.2",
     "tsx": "^4.7.1",
     "typescript": "^5.4.3",
